requirements.txt
absl-py==2.0.0
addict==2.4.0
ansi2html==1.8.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==23.11.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
ccimport==0.4.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.1.1
cumm-cu113==0.2.8
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
einops==0.7.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.0
fire==0.5.0
flake8==5.0.4
flash-attn==0.2.2
Flask==3.0.0
fonttools==4.45.0
fqdn==1.5.1
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.5.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
lark==1.1.8
llvmlite==0.36.0
loguru==0.7.2
lyft-dataset-sdk==0.0.8
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mmcls==0.25.0
mmcv-full @ file:///mnt/mmcv_full-1.6.0-cp38-cp38-manylinux1_x86_64.whl#sha256=7d0538129a125054818b2f9fa3194dd062c6e77802e85ce499609bca57bf5257
-e git+https://github.com/open-mmlab/mmdetection.git@1376e77e6ecbaad609f6003725158de24ed42e84#egg=mmdet
-e git+https://github.com/open-mmlab/mmdetection3d.git@fcb4545ce719ac121348cab59bac9b69dd1b1b59#egg=mmdet3d
mmengine==0.9.0
mmsegmentation==0.29.1
mpi4py @ file:///croot/mpi4py_1671223370575/work
msgpack==1.0.8
msgpack-numpy==0.4.8
multimethod==1.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==2.2
ninja==1.11.1.1
notebook==7.1.1
notebook_shim==0.2.4
numba==0.53.0
numpy==1.23.4
nuscenes-devkit==1.1.11
oauthlib==3.2.2
open3d==0.11.0
opencv-python==4.8.1.78
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.11.2
pccm==0.4.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.3.0
plyfile==1.0.2
portalocker==2.8.2
prettytable==3.9.0
prometheus_client==0.20.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.7
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.17.1
pyparsing==3.1.1
pyquaternion==0.9.9
pytest==7.4.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
RoiAwarePoolHelper==1.0
rpds-py==0.13.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.10.0
Send2Trash==1.8.2
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spconv-cu113==2.1.21
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorpack==0.11
termcolor==2.3.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.11.0+cu113
torch-scatter @ file:///root/CMT/torch_scatter-2.0.9-cp38-cp38-linux_x86_64.whl#sha256=ae2b2e666c8fa8383d639694463fff53f1e06ba66bdaa6b6cdd8a4aebe3f025a
torchaudio==0.11.0+cu113
torchpack==0.3.1
torchvision==0.12.0+cu113
tornado==6.4
tqdm==4.66.1
traitlets==5.13.0
trimesh==2.35.39
types-python-dateutil==2.8.19.20240311
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.11
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yapf==0.40.1
zipp==3.17.0
